src/main/java/ru/project/service/UserService.java
package ru.project.service;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import org.springframework.util.Assert;
import ru.project.model.User;
import ru.project.repository.UserRepository;

import static ru.project.util.ValidationUtil.checkNotFoundWithId;

@Service
public class UserService {
    private final UserRepository repository;

    @Autowired
    public UserService(UserRepository repository) {
        this.repository = repository;
    }

    public User create(User user) {
        return repository.save(user);
    }

    public void update(User user) {
        Assert.notNull(user, "user must not be null");
        checkNotFoundWithId(repository.save(user), user.getId());
    }

    public User get(int id) {
        return checkNotFoundWithId(repository.get(id), id);
    }

    public User getWithRestaurants(int id) {
        return checkNotFoundWithId(repository.getWithRestaurants(id), id);
    }

    public void delete(int id) {
        checkNotFoundWithId(repository.delete(id), id);
    }
}
